refactored connection handling class and methods
