+++ modules/parse/gocdb_downtimes.py
@@ -3,7 +3,7 @@ import xml.dom.minidom
 
 from xml.parsers.expat import ExpatError
 from argo_egi_connectors.helpers import module_class_name
-from argo_egi_connectors.input import ConnectorError
+from argo_egi_connectors.io.connection import ConnectorError
 
 
 class ParseDowntimes(object):
